OCPBUGS-44576: Resolve min-length cycle in `receiver-forms`
