+++ frontend/public/components/monitoring/receiver-forms/alert-manager-receiver-forms.tsx
@@ -4,11 +4,11 @@ import * as _ from 'lodash-es';
 import { Helmet } from 'react-helmet';
 import { useTranslation } from 'react-i18next';
 import { useParams, useNavigate } from 'react-router-dom-v5-compat';
-import { ActionGroup, Alert, Button, Tooltip } from '@patternfly/react-core';
+import { ActionGroup, Alert, Button } from '@patternfly/react-core';
 import { safeLoad } from 'js-yaml';
 import * as classNames from 'classnames';
 
-import { BlueInfoCircleIcon, APIError } from '@console/shared';
+import { APIError } from '@console/shared';
 import { ButtonBar } from '../../utils/button-bar';
 import { Dropdown } from '../../utils/dropdown';
 import { Firehose } from '../../utils/firehose';
@@ -470,64 +470,6 @@ const ReceiverBaseForm: React.FC<ReceiverBaseFormProps> = ({
   );
 };
 
-export const SaveAsDefaultCheckbox: React.FC<SaveAsDefaultCheckboxProps> = ({
-  formField,
-  disabled,
-  label,
-  formValues,
-  dispatchFormChange,
-  tooltip,
-}) => {
-  const saveAsDefaultLabelClass = classNames('checkbox', { 'co-no-bold': disabled });
-  return (
-    <label className={saveAsDefaultLabelClass} htmlFor={formField}>
-      <input
-        type="checkbox"
-        id={formField}
-        data-test-id="save-as-default"
-        onChange={(e) =>
-          dispatchFormChange({
-            type: 'setFormValues',
-            payload: { [formField]: e.target.checked },
-          })
-        }
-        checked={formValues[formField]}
-        aria-checked={formValues[formField]}
-        disabled={disabled}
-        aria-disabled={disabled}
-      />
-      <span className="co-alert-manager-config__save-as-default-label">{label}</span>
-      <Tooltip content={<p>{tooltip}</p>}>
-        <BlueInfoCircleIcon />
-      </Tooltip>
-    </label>
-  );
-};
-
-export const SendResolvedAlertsCheckbox = ({ formField, formValues, dispatchFormChange }) => {
-  const { t } = useTranslation();
-  return (
-    <div className="checkbox">
-      <label className="control-label" htmlFor={formField}>
-        <input
-          type="checkbox"
-          id={formField}
-          data-test-id="send-resolved-alerts"
-          onChange={(e) =>
-            dispatchFormChange({
-              type: 'setFormValues',
-              payload: { [formField]: e.target.checked },
-            })
-          }
-          checked={formValues[formField]}
-          aria-checked={formValues[formField]}
-        />
-        {t('public~Send resolved alerts to this receiver?')}
-      </label>
-    </div>
-  );
-};
-
 const ReceiverWrapper: React.FC<ReceiverFormsWrapperProps> = React.memo(({ obj, ...props }) => {
   const { alertManagerBaseURL } = window.SERVER_FLAGS;
   const [alertmanagerGlobals, setAlertmanagerGlobals] = React.useState();
@@ -640,21 +582,6 @@ type FormState = {
   receiverType: string;
   routeLabels: any[];
   [key: string]: string | any[] | any;
-};
-
-export type FormProps = {
-  globals: { [key: string]: any };
-  formValues: { [key: string]: any };
-  dispatchFormChange: Function;
-};
-
-type SaveAsDefaultCheckboxProps = {
-  formField: string;
-  disabled: boolean;
-  label: string;
-  formValues: { [key: string]: any };
-  dispatchFormChange: Function;
-  tooltip: string;
 };
 
 type ReceiverInfoTipProps = {
