+++ frontend/public/locales/en/public.json
@@ -1066,7 +1066,6 @@
   "Receiver name": "Receiver name",
   "A receiver with that name already exists.": "A receiver with that name already exists.",
   "Receiver type": "Receiver type",
-  "Send resolved alerts to this receiver?": "Send resolved alerts to this receiver?",
   "Alertmanager globals": "Alertmanager globals",
   "To address": "To address",
   "The email address to send notifications to.": "The email address to send notifications to.",
@@ -1104,6 +1103,7 @@
   "Incident details": "Incident details",
   "Description of the incident.": "Description of the incident.",
   "Severity of the incident.": "Severity of the incident.",
+  "Send resolved alerts to this receiver?": "Send resolved alerts to this receiver?",
   "Firing alerts with labels that match all of these <2></2> will be sent to this receiver.": "Firing alerts with labels that match all of these <2></2> will be sent to this receiver.",
   "matchers": "matchers",
   "Matcher": "Matcher",
